requirements.txt
django-ninja
psycopg2
